requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
sqlalchemy==2.0.48
psycopg2==2.9.11
pydantic==2.12.5
python-jose==3.5.0
cryptography==46.0.6
